move `flush_alt_blocks`
